+++ cogs/fun.py
@@ -12,7 +12,7 @@ import os
 from glob import glob
 import random
 from utils.CampBuddyMakov import CampBuddyMakov
-import utils.yuuto as yuuto
+import logging
 
 
 def find_next_user(first: discord.Member, seconds: List[discord.Member]) -> Optional[discord.Member]:
@@ -57,7 +57,7 @@ def check_rig(id1, id2):
 quote_enabled = bool(glob('./assets/quote/*.txt'))
 
 if not quote_enabled:
-    yuuto.logger.warn('No quotes found, disabling command')
+    logging.warn('No quotes found, disabling command')
 
 
 class Fun(commands.Cog):
